use port 443 instead
port 80 works, but actually only convenient for http, not https.

http -> port 80
https -> port 443

https://github.com/elixir-lang/elixir/blob/1a1eda52b46e37b0c77835e374c16ec64bb5c638/lib/elixir/src/elixir.erl#L48

Assuming that you're using https, calling `HelloElixir.Endpoint.url()`
returns `https://[slug].fly.dev:80` if the port is 80. Using port 443
returns `https://[slug].fly.dev` without the port in the url.

This is because the port is ignored if the scheme matches the default port.
In this case, https matches 443, so it's omitted.

The implementation actually happens in Elixir's URI module:
https://github.com/elixir-lang/elixir/blob/fb729784e5504f499b0179c4f154e73ea2a6f216/lib/elixir/lib/uri.ex#L948